requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
pip==22.1.1
platformdirs==2.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
setuptools==65.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
wheel==0.37.1
whitenoise==6.2.0
zope.interface==5.4.0
